requirements.txt
requests
fastapi
uvicorn
python-dotenv
bs4
redis